Additional Exceptions and Logging
Every Method in cl_messenger and cl_alarm are now logging the coding
info. There is a helper method currently in pi_ager_logging. Maybe it is
better to use a own helper class for this.
Currently the logger output in this two classes are set to INFO for
better testing. This will be set to DEBUG soon.